Handle Exceptions Better
The "parent deferrred" that tracks the process of the whole iteration
process now gets rejected (rather than resolved) immediately and its
promise returned if the collection passed is empty. THIS IS A BREAKING
CHANGE that alters expected behavior sufficiently for a major version
bump.

The parent deferred is also rejected if the passed callback is falsy or
not a function.

All completion callbacks now return both the collection passed to
jQuery.deferredEach() and a message indicating why the operation was
or wasn't successful.

README updated with all of the above info.